+++ message_ix_models/model/transport/operator.py
@@ -62,11 +62,13 @@ __all__ = [
     "distance_ldv",
     "distance_nonldv",
     "dummy_prices",
+    "duration_period",
     "extend_y",
     "factor_fv",
     "factor_input",
     "factor_pdt",
     "groups_iea_eweb",
+    "groups_y_annual",
     "iea_eei_fv",
     "indexers_n_cd",
     "indexers_usage",
@@ -80,6 +82,7 @@ __all__ = [
     "price_units",
     "quantity_from_config",
     "relabel2",
+    "sales_fraction_annual",
     "share_weight",
     "smooth",
     "transport_check",
@@ -317,22 +320,37 @@ def dummy_prices(gdp: "AnyQuantity") -> "AnyQuantity":
     )
 
 
-def extend_y(qty: "AnyQuantity", y: List[int]) -> "AnyQuantity":
+def duration_period(info: "ScenarioInfo") -> "AnyQuantity":
+    """Convert ``duration_period`` parameter data to :class:`.Quantity`.
+
+    .. todo:: Move to a more general module/location.
+    """
+    from genno.operator import unique_units_from_dim
+
+    return genno.Quantity(
+        info.par["duration_period"]
+        .replace("y", "year")
+        .rename(columns={"year": "y"})
+        .set_index(["y", "unit"])["value"]
+    ).pipe(unique_units_from_dim, "unit")
+
+
+def extend_y(qty: "AnyQuantity", y: List[int], *, dim: str = "y") -> "AnyQuantity":
     """Extend `qty` along the "y" dimension to cover `y`."""
     y_ = set(y)
 
     # Subset of `y` appearing in `qty`
-    y_qty = sorted(set(qty.to_series().reset_index()["y"].unique()) & y_)
+    y_qty = sorted(set(qty.to_series().reset_index()[dim].unique()) & y_)
     # Subset of `target_years` to fill forward from the last period in `qty`
     y_to_fill = sorted(filter(partial(lt, y_qty[-1]), y_))
 
-    log.info(f"{qty.name}: extend from {y_qty[-1]} → {y_to_fill}")
+    log.info(f"{qty.name}: extend '{dim}' from {y_qty[-1]} → {y_to_fill}")
 
     # Map existing labels to themselves, and missing labels to the last existing one
     y_map = [(y, y) for y in y_qty] + [(y_qty[-1], y) for y in y_to_fill]
     # - Forward-fill *within* `qty` existing values.
     # - Use message_ix_models MappingAdapter to do the rest.
-    return MappingAdapter({"y": y_map})(qty.ffill("y"))  # type: ignore [attr-defined]
+    return MappingAdapter({dim: y_map})(qty.ffill(dim))  # type: ignore [attr-defined]
 
 
 def factor_fv(n: List[str], y: List[int], config: dict) -> "AnyQuantity":
@@ -531,6 +549,17 @@ def groups_iea_eweb(technologies: List[Code]) -> Tuple[Groups, Groups, Dict]:
     g2["t"] = xr.DataArray(g2.pop("t"), coords=[("t_new", g2.pop("t_new"))])
 
     return g0, g1, g2
+
+
+def groups_y_annual(duration_period: "AnyQuantity") -> "AnyQuantity":
+    """Return a list of groupers for aggregating annual data to MESSAGE periods.
+
+    .. todo:: Move to a more general module/location.
+    """
+    result = {}
+    for (period,), duration in duration_period.to_series().items():
+        result[period] = list(range(period + 1 - int(duration), period + 1))
+    return dict(y=result)
 
 
 def input_commodity_level(t: List[Code], default_level=None) -> "AnyQuantity":
@@ -782,6 +811,14 @@ def relabel2(qty: "AnyQuantity", new_dims: dict):
         result = select(result, selectors)
 
     return result
+
+
+def sales_fraction_annual(y0: int, age: int) -> "AnyQuantity":
+    """Return fractions of current stock that should be sold in prior years."""
+    N_y = 2 * int(age)  # Number of periods
+    return genno.Quantity(
+        [1.0 / N_y] * N_y, coords={"y": range(y0 + 1 - N_y, y0 + 1)}, units=""
+    )
 
 
 def share_weight(